+++ core/build.gradle
@@ -1,6 +1,4 @@
 dependencies {
-    compile project(":metrics")
-
     compile group: 'com.google.code.gson', name: 'gson', version: '2.8.5'
 }
