--- metrics/src/main/kotlin/br/com/guiabolso/metrics/async/DefaultAsyncExecutor.kt
@@ -1,28 +1,0 @@
-package br.com.guiabolso.metrics.async
-
-import br.com.guiabolso.metrics.engine.MetricReporterEngine
-import java.util.concurrent.Callable
-import java.util.concurrent.ExecutorService
-import java.util.concurrent.Future
-
-class DefaultAsyncExecutor : AsyncExecutor {
-
-    override fun <T> executeAsync(engine: MetricReporterEngine<*>, executor: ExecutorService, task: () -> T): Future<T> {
-        val context = engine.extractContext()!!
-        return executor.submit((Callable {
-            engine.withContext(context).use {
-                task()
-            }
-        }))
-    }
-
-    override fun <T> executeAsync(engine: MetricReporterEngine<*>, executor: ExecutorService, task: Callable<T>): Future<T> {
-        val context = engine.extractContext()!!
-        return executor.submit((Callable {
-            engine.withContext(context).use {
-                task.call()
-            }
-        }))
-    }
-
-}
