--- metrics/src/main/kotlin/br/com/guiabolso/metrics/engine/MetricReporterEngine.kt
@@ -1,82 +1,0 @@
-package br.com.guiabolso.metrics.engine
-
-import java.io.Closeable
-
-interface MetricReporterEngine<C> {
-
-    /**
-     * Defines the operation name.
-     *
-     * @param name operation name.
-     * @since 2.0.0
-     */
-    fun setOperationName(name: String)
-
-    /**
-     * Add a key/value pair to the current metric operation. These should be reported in errors and tracings.
-     *
-     * @param key Custom parameter key.
-     * @param value Custom parameter value.
-     * @since 2.0.0
-     */
-    fun addProperty(key: String, value: String?)
-
-    /**
-     * Add a key/value pair to the current metric operation. These should be reported in errors and tracings.
-     *
-     * @param key Custom parameter key.
-     * @param value Custom parameter value.
-     * @since 2.0.0
-     */
-    fun addProperty(key: String, value: Number?)
-
-    /**
-     * Add a key/value pair to the current metric operation. These should be reported in errors and tracings.
-     *
-     * @param key Custom parameter key.
-     * @param value Custom parameter value.
-     * @since 2.0.0
-     */
-    fun addProperty(key: String, value: Boolean?)
-
-    /**
-     * Notice an error and report it to the metric reporter.
-     *
-     * @param exception The exception to be reported.
-     * @param expected true if this error is expected, false otherwise.
-     * @since 2.0.0
-     */
-    fun notifyError(exception: Throwable, expected: Boolean)
-
-    /**
-     * Notice an error and report it to the metric reporter.
-     *
-     * @param message Error message.
-     * @param params Custom parameters to include in the traced error. May be null.
-     * @param expected true if this error is expected, false otherwise.
-     * @since 2.0.0
-     */
-    fun notifyError(message: String, params: Map<String, String?>, expected: Boolean)
-
-    /**
-     * Extract the current trace context
-     */
-    fun extractContext(): C
-
-    /**
-     * Uses the current trace context. After the execution the the context must be closed.
-     * Its safer to use {@link #withContext(context: C, func: () -> Any) withContext}
-     */
-    fun withContext(context: Any): Closeable
-
-    /**
-     * Uses the current trace context
-     */
-    fun withContext(context: C, func: () -> Any)
-
-    /**
-     * Cleans the metric reporter state.
-     */
-    fun clear()
-
-}
