--- metrics/src/main/kotlin/br/com/guiabolso/metrics/engine/slf4j/MDCCloseable.kt
@@ -1,10 +1,0 @@
-package br.com.guiabolso.metrics.engine.slf4j
-
-import org.slf4j.MDC
-import java.io.Closeable
-
-object MDCCloseable : Closeable {
-    override fun close() {
-        MDC.clear()
-    }
-}
