--- metrics/src/main/kotlin/br/com/guiabolso/metrics/utils/ClasspathUtils.kt
@@ -1,16 +1,0 @@
-package br.com.guiabolso.metrics.utils
-
-object ClasspathUtils {
-
-    fun isNewRelicPresent() = isClassPresent("com.newrelic.api.agent.NewRelic")
-
-    fun isDatadogPresent() = isClassPresent("datadog.opentracing.DDTracer")
-
-    private fun isClassPresent(name: String) = try {
-        Class.forName(name)
-        true
-    } catch (_: ClassNotFoundException) {
-        false
-    }
-
-}
