+++ server/src/main/kotlin/br/com/guiabolso/events/server/exception/EventExceptionHandler.kt
@@ -2,7 +2,7 @@ package br.com.guiabolso.events.server.exception
 
 import br.com.guiabolso.events.model.RequestEvent
 import br.com.guiabolso.events.model.ResponseEvent
-import br.com.guiabolso.metrics.MetricReporter
+import br.com.guiabolso.tracing.MetricReporter
 
 interface EventExceptionHandler<in T : Throwable> {
