+++ server/src/main/kotlin/br/com/guiabolso/events/server/exception/LambdaEventExceptionHandler.kt
@@ -2,7 +2,7 @@ package br.com.guiabolso.events.server.exception
 
 import br.com.guiabolso.events.model.RequestEvent
 import br.com.guiabolso.events.model.ResponseEvent
-import br.com.guiabolso.metrics.MetricReporter
+import br.com.guiabolso.tracing.MetricReporter
 
 class LambdaEventExceptionHandler<in T : Throwable>(private val func: (T, RequestEvent, MetricReporter) -> ResponseEvent) : EventExceptionHandler<T> {
