+++ settings.gradle
@@ -1,2 +1,2 @@
 rootProject.name = 'events-protocol'
-include 'metrics', 'core', 'server', 'client'
+include 'tracing', 'core', 'server', 'client'
