+++ tracing/build.gradle
@@ -1,0 +1,32 @@
+dependencies {
+    compile group: 'org.slf4j', name: 'slf4j-api', version: '1.7.25'
+
+    //NewRelic
+    compileOnly group: 'com.newrelic.agent.java', name: 'newrelic-api', version: '4.4.0'
+
+    //Datadog
+    compileOnly group: 'io.opentracing', name: 'opentracing-api', version: "0.31.0"
+    compileOnly group: 'io.opentracing', name: 'opentracing-util', version: "0.31.0"
+    compileOnly group: 'com.datadoghq', name: 'dd-trace-ot', version: '0.12.0'
+    compileOnly group: 'com.datadoghq', name: 'dd-trace-api', version: '0.12.0'
+
+    compile "ch.qos.logback:logback-classic:1.2.3"
+    testCompile group: 'io.opentracing', name: 'opentracing-api', version: "0.31.0"
+    testCompile group: 'io.opentracing', name: 'opentracing-util', version: "0.31.0"
+    testCompile group: 'com.datadoghq', name: 'dd-trace-ot', version: '0.12.0'
+    testCompile group: 'com.datadoghq', name: 'dd-trace-api', version: '0.12.0'
+}
+
+apply plugin: 'com.novoda.bintray-release'
+
+publish {
+    artifactId = 'tracing'
+    autoPublish = true
+    desc = 'Tracing library'
+    userOrg = 'guiabolso'
+    groupId = 'br.com.guiabolso'
+    licences = ['APACHE-2.0']
+    publishVersion = parent.ext.publish_version
+    uploadName = 'Metrics'
+    website = 'https://github.com/GuiaBolso/events-protocol'
+}
