+++ tracing/src/main/kotlin/br/com/guiabolso/tracing/engine/newrelic/CloseableToken.kt
@@ -1,0 +1,10 @@
+package br.com.guiabolso.tracing.engine.newrelic
+
+import com.newrelic.api.agent.Token
+import java.io.Closeable
+
+data class CloseableToken(private val token: Token) : Closeable {
+    override fun close() {
+        token.expire()
+    }
+}
