+++ tracing/src/main/kotlin/br/com/guiabolso/tracing/engine/slf4j/MDCCloseable.kt
@@ -1,0 +1,10 @@
+package br.com.guiabolso.tracing.engine.slf4j
+
+import org.slf4j.MDC
+import java.io.Closeable
+
+object MDCCloseable : Closeable {
+    override fun close() {
+        MDC.clear()
+    }
+}
