+++ tracing/src/main/kotlin/br/com/guiabolso/tracing/utils/ClassPathUtils.kt
@@ -1,0 +1,16 @@
+package br.com.guiabolso.tracing.utils
+
+object ClassPathUtils {
+
+    fun isNewRelicPresent() = isClassPresent("com.newrelic.api.agent.NewRelic")
+
+    fun isDatadogPresent() = isClassPresent("datadog.opentracing.DDTracer")
+
+    private fun isClassPresent(name: String) = try {
+        Class.forName(name)
+        true
+    } catch (_: ClassNotFoundException) {
+        false
+    }
+
+}
