+++ tracing/src/main/kotlin/br/com/guiabolso/tracing/utils/ExceptionUtils.kt
@@ -1,0 +1,16 @@
+package br.com.guiabolso.tracing.utils
+
+import java.io.PrintWriter
+import java.io.StringWriter
+
+object ExceptionUtils {
+
+    @JvmStatic
+    fun getStackTrace(throwable: Throwable): String {
+        val sw = StringWriter()
+        val pw = PrintWriter(sw, true)
+        throwable.printStackTrace(pw)
+        return sw.buffer.toString()
+    }
+
+}
